update external module
