+++ src/index.html
@@ -24,6 +24,7 @@
               placeholder="Add To Your List..."
               type="text"
               class="full-width wrapper input"
+              id="todo-task"
             />
             <i class="fa fa-reply wrapper"></i>
           </form>
